[REFACTOR] zzim 도메인 분리
