--- src/main/java/com/spoony/spoony_server/domain/post/dto/response/ZzimFocusListResponseDTO.java
@@ -1,6 +1,0 @@
-package com.spoony.spoony_server.domain.post.dto.response;
-
-import java.util.List;
-
-public record ZzimFocusListResponseDTO(List<ZzimFocusResponseDTO> zzimFocusResponseList){
-}
