+++ src/main/java/com/spoony/spoony_server/domain/zzim/dto/request/ZzimPostAddRequestDTO.java
@@ -1,4 +1,4 @@
-package com.spoony.spoony_server.domain.post.dto.request;
+package com.spoony.spoony_server.domain.zzim.dto.request;
 
 public record ZzimPostAddRequestDTO(Long userId, Long postId) {
 }
