+++ src/main/java/com/spoony/spoony_server/domain/zzim/dto/response/ZzimCardResponseDTO.java
@@ -1,4 +1,6 @@
-package com.spoony.spoony_server.domain.post.dto.response;
+package com.spoony.spoony_server.domain.zzim.dto.response;
+
+import com.spoony.spoony_server.domain.post.dto.response.CategoryColorResponseDTO;
 
 public record ZzimCardResponseDTO(Long placeId,
                                   String placeName,
