+++ src/main/java/com/spoony/spoony_server/domain/zzim/dto/response/ZzimFocusListResponseDTO.java
@@ -1,0 +1,6 @@
+package com.spoony.spoony_server.domain.zzim.dto.response;
+
+import java.util.List;
+
+public record ZzimFocusListResponseDTO(List<ZzimFocusResponseDTO> zzimFocusResponseList) {
+}
