+++ src/main/java/com/spoony/spoony_server/domain/zzim/service/ZzimPostService.java
@@ -1,4 +1,4 @@
-package com.spoony.spoony_server.domain.post.service;
+package com.spoony.spoony_server.domain.zzim.service;
 
 import com.spoony.spoony_server.common.dto.ResponseDTO;
 import com.spoony.spoony_server.common.exception.BusinessException;
@@ -9,18 +9,22 @@ import com.spoony.spoony_server.domain.location.entity.LocationEntity;
 import com.spoony.spoony_server.domain.location.repository.LocationRepository;
 import com.spoony.spoony_server.domain.place.entity.PlaceEntity;
 import com.spoony.spoony_server.domain.place.repository.PlaceRepository;
-import com.spoony.spoony_server.domain.post.dto.request.ZzimPostAddRequestDTO;
-import com.spoony.spoony_server.domain.post.dto.response.*;
+import com.spoony.spoony_server.domain.post.dto.response.CategoryColorResponseDTO;
 import com.spoony.spoony_server.domain.post.entity.PhotoEntity;
 import com.spoony.spoony_server.domain.post.entity.PostCategoryEntity;
 import com.spoony.spoony_server.domain.post.entity.PostEntity;
-import com.spoony.spoony_server.domain.post.entity.ZzimPostEntity;
 import com.spoony.spoony_server.domain.post.repository.PhotoRepository;
 import com.spoony.spoony_server.domain.post.repository.PostCategoryRepository;
 import com.spoony.spoony_server.domain.post.repository.PostRepository;
-import com.spoony.spoony_server.domain.post.repository.ZzimPostRepository;
 import com.spoony.spoony_server.domain.user.entity.UserEntity;
 import com.spoony.spoony_server.domain.user.repository.UserRepository;
+import com.spoony.spoony_server.domain.zzim.dto.request.ZzimPostAddRequestDTO;
+import com.spoony.spoony_server.domain.zzim.dto.response.ZzimCardListResponseDTO;
+import com.spoony.spoony_server.domain.zzim.dto.response.ZzimCardResponseDTO;
+import com.spoony.spoony_server.domain.zzim.dto.response.ZzimFocusListResponseDTO;
+import com.spoony.spoony_server.domain.zzim.dto.response.ZzimFocusResponseDTO;
+import com.spoony.spoony_server.domain.zzim.entity.ZzimPostEntity;
+import com.spoony.spoony_server.domain.zzim.repository.ZzimPostRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
